chore(be): 필요 없거나 잘못된 의미 수정

+++ backend/src/main/java/com/deliveranything/domain/order/service/PaymentOrderService.java
@@ -27,7 +27,7 @@ public class PaymentOrderService {
     order.isPayable();
 
     eventPublisher.publishEvent(new OrderPaymentRequestedEvent(order.getId(), paymentKey,
-        merchantUid, order.getTotalPrice().longValue()));
+        merchantUid, order.getTotalPrice()));
 
     return OrderResponse.from(order);
   }
